Feature: File permission checker
Added a checker to see if files have permissions before trying to execute to them because it was leaking a lot when it went through and open checked for the perms. Now behaves more like bash I think.
Started working on the tests and started giving notations when needed.
Might be relevant to check if the directory we are trying to create files to has write permissions, otherwise there might be some errors